requirements.txt
# ==========================
# Core deep learning
# ==========================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ==========================
# EEG + signal processing
# ==========================
mne>=1.5.0
scipy>=1.10.0
numpy>=1.24.0

# ==========================
# ML & evaluation
# ==========================
scikit-learn>=1.3.0

# ==========================
# Visualization
# ==========================
matplotlib>=3.7.0
seaborn>=0.12.2

# ==========================
# Utils
# ==========================
pyyaml>=6.0
tensorboard>=2.12.0
tqdm>=4.65.0
